Add DictionaryBuilder trait.
