+++ lindera-core/src/lib.rs
@@ -1,8 +1,11 @@
 pub mod character_definition;
 pub mod connection;
+pub mod dictionary_builder;
 pub mod error;
 pub mod prefix_dict;
 pub mod unknown_dictionary;
+pub mod user_dictionary;
+pub mod util;
 pub mod viterbi;
 pub mod word_entry;
