+++ lindera-core/src/user_dictionary.rs
@@ -1,0 +1,7 @@
+use crate::prefix_dict::PrefixDict;
+
+pub struct UserDictionary {
+    pub dict: PrefixDict<Vec<u8>>,
+    pub words_idx_data: Vec<u8>,
+    pub words_data: Vec<u8>,
+}
