+++ lindera-ipadic-builder/src/main.rs
@@ -2,8 +2,9 @@ use std::path::Path;
 
 use clap::{crate_authors, crate_description, crate_name, crate_version, App, AppSettings, Arg};
 
+use lindera_core::dictionary_builder::DictionaryBuilder;
 use lindera_core::LinderaResult;
-use lindera_ipadic_builder::builder::build;
+use lindera_ipadic_builder::ipadic_builder::IpadicBuilder;
 
 fn main() -> LinderaResult<()> {
     let app = App::new(crate_name!())
@@ -34,7 +35,9 @@ fn main() -> LinderaResult<()> {
     let input_dir = Path::new(matches.value_of("INPUT_DIR").unwrap()).to_path_buf();
     let output_dir = Path::new(matches.value_of("OUTPUT_DIR").unwrap()).to_path_buf();
 
-    match build(&input_dir, &output_dir) {
+    let builder = IpadicBuilder::new();
+
+    match builder.build_dictionary(&input_dir, &output_dir) {
         Ok(()) => println!("{}", "done"),
         Err(msg) => println!("{}", msg),
     }
