+++ lindera-ipadic/Cargo.toml
@@ -30,4 +30,5 @@ tar = "0.4"
 encoding = "0.2"
 tokio = { version = "1.9", features = ["full"] }
 
+lindera-core = { version = "0.7", path = "../lindera-core" }
 lindera-ipadic-builder = { version = "0.7", path = "../lindera-ipadic-builder" }
